+++ src/clj_cbor/codec.clj
@@ -533,7 +533,7 @@
   [decoder ^DataInputStream input info]
   (let [tag (header/read-code input info)
         value (read-value decoder input)]
-    (if (= tag (:set-tag decoder))
+    (if (= tag data/set-tag)
       (read-set decoder value)
       (try
         (if-let [handler ((:read-handlers decoder) tag)]
@@ -702,7 +702,7 @@
     (seq? x)    (write-array codec out x)
     (vector? x) (write-array codec out x)
     (map? x)    (write-map codec out x)
-    (set? x)    (write-set codec out (:set-tag codec) x)
+    (set? x)    (write-set codec out data/set-tag x)
     :else       nil))
 
 
@@ -910,7 +910,7 @@
 ;; ## Codec Record
 
 (defrecord CBORCodec
-  [dispatch write-handlers read-handlers set-tag]
+  [dispatch write-handlers read-handlers]
 
   Encoder
 
@@ -939,6 +939,5 @@
     {:dispatch class
      :write-handlers {}
      :read-handlers {}
-     :set-tag 258
      :canonical false
      :strict false}))